Resolved issue #463 - Compile Error:
The compiler was running out of memory when constructing precompiled headers.
After setting the '/Zm' flag in Cmake the compiler should now automatically
be instructed to allocate memory when constructing precompiled headers.

This fix worked for me when setting the flag to 200(150MB) and 1000(750MB),
but should be further validated and tested.