Resolved issue #108 - Spell channeling and auras:
Spells such as Mind Flay and Drain Soul could not have be channeled by more then one player
per single target. If the second player decided to channel the same spell at the same target,
the channeling event for the first player would be canceled/interrupted, and the aura would be refreshed.

It took hours of trial and error to try to figure out what was going on 'AddAura', and after
rewriting the section that handles aura stacking, LordUsagi and myself finally made it work.

There is a small visual bug with aura icons that remains unsolved, and it's uncertain
if it was present there before or was it recent changes in code that triggered this odd behavior.
Other then that everything seems to be working in perfect order!